Getting close if not there with tokens and tests.
