+++ README.md
@@ -1,24 +1,11 @@
-# README
+# Kam's Tikanga(knowledge)
 
-This README would normally document whatever steps are necessary to get the
-application up and running.
+Various applications 
 
-Things you may want to cover:
+##### Prerequisites
 
-* Ruby version
+The setup steps expect the following tools are installed on the system.
 
-* System dependencies
-
-* Configuration
-
-* Database creation
-
-* Database initialization
-
-* How to run the test suite
-
-* Services (job queues, cache servers, search engines, etc.)
-
-* Deployment instructions
-
-* ...
+- Github
+- Ruby [3.3.0](https://www.ruby-lang.org/en/news/2023/12/25/ruby-3-3-0-released/)
+- Rails [8.0.2](https://rubyonrails.org/2025/3/12/Rails-Version-8-0-2-has-been-released)
