+++ app/controllers/mobile_forecast_controller.rb
@@ -15,17 +15,19 @@ class MobileForecastController < ApplicationController
     render partial: "full", locals: { location_name: location_context.location_name, erred: erred, forecasts: forecasts }
   end
 
+  def geo_location
+    erred, locations, total = location_services(params[:location])
+    multi_locations(erred, locations, total)
+  end
+
+  private
+
   def full_location(locations)
     found_location = locations.first
     params.merge!(lat: found_location["geometry"]["lat"], long: found_location["geometry"]["lng"], location_name: found_location["formatted"])
     location_context = LocationContext.new(params)
     erred, forecasts = create_forecasts(latitude: location_context.latitude, longitude: location_context.longitude)
     render turbo_stream: turbo_stream.replace(FULL_TURBO, partial: "full", locals: { location_name: location_context.location_name, erred: erred, forecasts: forecasts })
-  end
-
-  def geo_location
-    erred, locations, total = location_services(params[:location])
-    multi_locations(erred, locations, total)
   end
 
   def multi_locations(erred, locations, total)
