+++ app/views/mobile_forecast/_geo_location.html.erb
@@ -1,10 +1,16 @@
 <turbo-frame id="location_response">
   <div class="table_mobile_geo_container">
+    <%
+      erred         ||= @erred
+      location      ||= @location
+      locations      ||= @locations
+      total      ||= @total
+    %>
     <div class="mobile_summary_div">
-      <% if @erred %>
-        <p class="fs-6">Location Response: <%= @locations %></p>
+      <% if erred %>
+        <p class="fs-6">Location Response: <%= locations %></p>
       <% else %>
-        <p class="fs-6">Location Response from: <%= @location %>, Total Locations: <%= @total %></p>
+        <p class="fs-6">Location Response from: <%= location %>, Total Locations: <%= total %></p>
         <table class="table">
           <thead>
           <tr>
@@ -12,13 +18,13 @@
           </tr>
           </thead>
           <tbody>
-          <% @locations.each do |location_hash| %>
+          <% locations.each do |location_hash| %>
             <tr>
               <td><p class="fs-8 geo_formatted"><%= location_hash["formatted"] %><br/>
                 <%= link_to 'Full Forecast', mobile_forecast_full_path(
                 lat: location_hash["geometry"]["lat"],
                 long: location_hash["geometry"]["lng"],
-                location: @location,
+                location: location,
                 location_name: location_hash["formatted"]
               ), data: { turbo_frame: 'full_response'} %> </p>
               </td>
