+++ test/controllers/concerns/authenticate_controller_test.rb
@@ -1,0 +1,16 @@
+require "test_helper"
+
+
+class AuthenticateController < ActionController::API
+  include AuthenticateApi
+
+  def test_action
+    auth_header = request.headers["Authorization"]
+  end
+end
+
+class AuthenticateControllerTest < ActionDispatch::IntegrationTest
+  test "will get index" do
+
+  end
+end
