+++ test/test_helper.rb
@@ -3,15 +3,16 @@ require_relative "../config/environment"
 require "rails/test_help"
 require "minitest/autorun"
 require "ostruct"
+require_relative "helpers/jwt_test_helper"
+require_relative "helpers/forecast_test_helper"
 
 module ActiveSupport
   class TestCase
     # Run tests in parallel with specified workers
     parallelize(workers: :number_of_processors)
-
+    include JwtTestHelper
+    include ForecastTestHelper
     # Setup all fixtures in test/fixtures/*.yml for all tests in alphabetical order.
     fixtures :all
-
-    # Add more helper methods to be used by all tests here...
   end
 end
